Initial Commit. Simple PHD mapping algorithm.
The PHD filter has been implemented for a 2D mapping application and an
appropiate simulation has been coded. The trajectory of the vehicle is
assumed known.
The specific version of the filter is the mapping part of the
Rao-Blackwellized PHD, which uses a gaussian mixture to represent the
estimate of the hidden state (i.e. the map).