Navigator trajectory draw and optimizations.
The drawing algorithm for the navigator particles has been changed to
follow the most likely particle at all times (jumping if necessary if such
a particle dies in the process).
A couple of simple optimizations has been performed on the Navigator
class to speed the update process. In particular, the most important ones
are using parallelism on the maps updates, avoiding a inversion on the
Mahalanobis function, avoiding copying all the trajectories on the Vehicle
class and pre-evaluating the exponential function on the Gaussian class.
Other minor modifications to style or encapsulating code into functions
were done.