Parameter selection on navigation.
Before, the vehicle particles on the navigator were perfect clones of the
original (real) vehicle. This is both unrealistic and impractical, since
the exact covariances aren't necessarily known and some of the
disadvantanges of the method may be mitigated by using different
parameters.
The gaussian render algorithm has been slightly improved. The border is no
longer completely linear with the weight (as it can become too intrusive
on the rest of the map visual info). Now there is a max cap when weight is
one. When this cap is reached, the border color goes black, when it gets
to two, it goes red.