Clutter simulation.
The vehicle simulator now creates spurious measurements following a
clutter density in a Poisson process.
The measurement space has been appropiately reduced to follow a physically
sensible model (and theoretically sound too): the range space is clipped
to a certain region. An example of this is the Kinect, which has a limited
measurement range.
In the case of monocular video, the range is effectively infinite but it
is not part of the measurement space as it is not measured.
The render procedure for the vehicle FOV has been adjusted to show this.
Though the model has a limited range (i.e. radial), the render model uses
a frustum with limited z-range (i.e. perpendicular distance). This is for
efficiency and simplicity reasons. The "sensed area" is not an exact
representation of the vehicle sensor but it is a good enough
visualization. The measure points are always correct, though.