Landmark-vehicle decoupling.
The vehicle model depended on the landmarks on each measure() call. This
was not appropiate to generalize to real input (as it is impossible to
give such landmarks to the sensor). The new way gets an internal reference
to the map. This way only the constructor requires access to the map, no
reference is needed when measuring. Ergo, the abstraction can be
accomplished using inheritance with a new class with an appropiate
constructor.