Most-probable-components particle weighting rough sketch.
Previously, the one-feature map trick had been used to find the
vehicle particle likelihood. Now a more sophisticated method is
implemented: the full map estimate is used and the likelihood of
every possible assignment should be computed. This is extremely
slow so a simple approximation is performed using Murty's algorithm.

In the special case where there are few map model components and
measurements all the possibilities are explored using a
lexicographical order.

To simplify the calculation, disconnected components are treated
separately.