Realtime option.
So far the system always reacts in realtime. If the operations take
too long, the vehicle moves around a lot in the corresponding step.
Now it can simulate a constant rate. This way the program is more
repeatable and the framework overhead can be bypassed (as in the real
vehicle an optimized version of the algorithm can be implemented).
Also, it gives perspective into the repercusions of the speed of
the algorithm.