Kinect sensor interface refined.
The interaction with the kinect has been refined and debugged.
OpenNI drivers have been added as well as a sample recording.
The GUI has been slightly altered to give more information: the
screen is split in two; the main scene screen panel shows the simulated
vehicle and map while the sidebar gives additional views.
In particular, when a Kinect sensor based vehicle is used, the original
depth bitmap is shown, along with the points of interest found on it.
Gaussian constructor changed to never produce NaN weights, as it
confuses the sorting algorithm.
IDisposable interface honored for a couple of fields in KinectVehicle.
Minor spelling and style corrections.