User modifiable mapping options.
The mode of the Navigator can now be set at runtime, changing
between doing slam and mapping. Along with it, reset functionality
has been added to simplify visualization and interacting with the
algorithm (positioning the vehicle somewhere and then starting
the process).

The visibility of some fields on the Navigator have been modified
to follow best practice guidelines.

Minor styling and grammar corrections have been added.