Added simulation tests.
Unit tests have been added for Simulation and related classes
(Explorer + Navigator). This tests are a rough starting point
and need to be refined and extended.
In the process some additional debug functionality has been added
to each class as well as minor behavior modifications.

The vehicle State property now copies by value instead of by reference.
The orientation motion covariance has been reduced to a more realistic
value. This allows to see the filter in working conditions more closely.
A small debug snippet has been added to the Navigator Update method to
allow the coder to force one particle to be the real vehicle (then
it can be asserted that it should be the best particle).
In the borderline case where no particle has any weight, the update
normalization doesn't give NaN but instead leeaves the particles
untouched.