Particle depletion detection.
Previously, the vehicle particles were resampled on every single
slam step. This is both wasteful and kills probing particles too early,
effectively focusing closely on the most probable hypothesis and
giving no option to the possibility of having made a mistake
somewhere on the past.
The new behaviour checks if the number of effective particles
(particles with enough weight) is too low and only then do the
resampling.