Sensor vehicle keypoint extraction.
The keypoint extraction algorithm has been changed to the FAST corner
detection. This change was based on some testing made with different
options (SURF, FREAK, Harris, FAST and NARF). The FAST algorithm is
by far the fastest one; the others (specially SURF) are probably not
that suitable for real-time operation. The depth map has very little
texture information anyway and all of them work pretty much the same
in terms of quality.

To improve even further the speed of the system, the sensor input
can now be subsampled before it is processed. This means smaller
processing times but also focusing on big-picture feature points,
which helps the detector if it usually gives too many corner points.

To decouple the speed of the keypoint extractor and the SLAM solver
a cache system has been implemented. It allows preprocessing a number
of frames (if the input is a file) beforehand. This way, the times
can be measured separately. The mechanism is trivial to turn off.

The visual output has been simplified, using only one image with the
keypoints drawn on top. Also, some functions that were no longer used
have been removed.

Some documentation comments have been corrected or improved.

It is noted that the algorithm will probably work much better using
the color space instead of the depth space for keypoint localization
as this space contains rich texture information. Next commits shall
explore that direction.

The build script now only overwrites files from the /additional folder
if they are newer than the ones in /bin. This is specially important
now, as the new test files are heavy and copying them around is
wasteful.

Some example oni files have been added to the project but are too specific
and heavy to put on the repo. They have been added to the .gitignore
file.
Preparing for the next stage, *.in files have also been ignored,
since they will be autogenerated.