Groundtruth to odometry.
A python script that takes the groundtruth movement of the vehicle
and generates an appropiate (perfect) odometry for the simulation has
been added. This way, the real examples can be used with a reasonable
amount of work. Otherwise, the number of particles needed explodes.
It uses an inverse update model.
A couple of groundtruth files were appended.