Modified vehicle simulation options.
Now the odometry input file can indicate the full 6DoF of the motion.
In particular lateral movement is controllable (though not in the
interactive input as it is too complex to control manually).

The odometry can be ignored. This may help when no meaningful odometry
is available.

The vehicle simulation can indicate if the nomovement position is
perfect or noise can still creep in when the vehicle is standing
still. This is important for handheld cameras.