Bugfix: odometry calculation.
The odometry calculation was incorrect because the conversion
between quaternions and yaw-pitch-roll was off. Also, the input
orientation quaternion were not necesssarily unit length and the
program assumed them to be. Now they are normalized first.
To check the correctness of the odometry output, the local deltas are
transformed back into a global estimate of the position and then
plotted against the original. It seems ok for the couple of
examples that were tried.