Feature extraction adjustments.
When extracting features from the depth map, the border may create
artificial keypoints. A minimum distance to it is now demanded for
a keypoint to be considered valid.

In the process of implementing this, a bug has been fixed: the
depth map has a multiplier of 5000 on the used dataset. The fix is
temporary though and should be extended to handle any multiplier.