Minor modifications to parameters.
The covariance matrices values, as well as the model max quantity
and simulated expected detection probability have been changed.

Note that if the detection probability is correctly estimated, i.e.
it is equal to the simulated one, the system works very well, but
if it is not, the system may show interesting phenomena like
double-weighted landmarks.