New Viewer mode.
Watching the process run on big maps or with lots of particles can
be tedious due to the framerate dropping considerably so a
recording option has been setup. After the user closes the main
application the system now saves all the relevant data in a more
organized fashion. A command line argument can now be used to
replay the experiment without the need to compute every update,
since they have already been computed.

This allows the user to see the flow of the process as if it was
realtime and to present results to others in a reliable manner.

Some restructuring in the classes and files has been performed
but are minor besides the new class Viewer and the modifications
required to make it play nicely with inheritance and avoid code
duplication.

An important change in behaviour in the interface is that the
command line argument -m is now used to input the map history file
instead of indicating if only-mapping was to be used; this feature
has been relegated to the argument -y.