Color kinect input basic functionality.
Kinect's color stream is now extracted from the device and used to
obtain the keypoints. They are remarkably better than those obtain
by only using the depth frame data.

There are a couple of things to iron out though: the framerate is
incorrect (x2) because as there are two streams at pretty much the
same data rate, every update only takes half the information (the
change in one of the streams but not the other), so it ends up
halving the speed; also, the sidebar should be recorded along with
the simulation.