Bugfix: framerate dropping.
To try and make the simulation go smoother, MonoGame drops render
calls if the update method takes too long. Though this is good for
general loops, it's not good to visualize the slow RFS simulation.

To counteract this, the system is now always "realtime", i.e.
updates as fast as possible. Nevertheless the realtime parameter
still does its thing: the simulated time can be non-realtime, even
if the render time always is. This means that if frames are too
cheap to compute the simulated may look too fast, but it should
look ok in the Viewer mode.