Viewer sidebar output.
The previously recorded sidebar video is now displayed when viewing
a simulation recording in sync with the rest of the output.

Note however that there is a bug (it seems to be in the library
code) so sometimes the video file doesn't load correctly. It may
have something to do with codecs or with memory management.