Linux port.
The package has been ported to linux, changing the file structure,
the dependencies and fixing inconsistent behaviour  as needed.

The most interesting change is that the libraries have a couple
of bugs specific to linux and they need som workarounds.

MonoGame for linux handes the closing of a window incorrectly.
This means that using the x button from the window manager
crashes the program. An alternative close key on the keyboard
has been set. The same bug also makes the Dispose pattern
to become broken, as it is not called timely and many objects
are invalid. The extent of the bug over the program is minor
so it's better to wait for a patch.

Accord.Extensions had a minor but important bug, where the video
writer and image reader miscalculated the image stride, thus
generating segmentation faults on the underlying opencv library.
This has been hotfixed library and an appropiate bug report has
been filed on the library project.

All libraries needed to be compiled against linux and linked to
the correct system libraries. It is noted that although very
practical in Windows, NuGet packages usually do not contain correct
binaries for linux. To change the behaviour of some of the packages
appropiate assembly .config files have been generated.

OpenNI2 Drivers and shared library have been compiled on linux. Note
that this version does not require a special driver.

Minor line endings normalizations.

An alternative solution file for linux has been generated.