Minor cleanup.
There was a couple of leftover code from previous implementations
and tests that now have been removed.

Tuple construciton now uses the handy Tuple.Create() alternative.

Waypoint "draw only when not too close" option has been removed.
It had been deactivated previously to facilitate the generation
of the Viewer code, but it really doesn't help much. The memory is
mostly used by maps, as they are much more heavyweight anyway and
having a variable timestep in the trajectories is just not worth
the hassle.