iSAM2 Graph SLAM basic blocks.
Any SLAM method should compare itself with the best state-of-art
alternatives. In this case, the graph-based family of algorithms
seem to be the best performing, so an iSAM2 solution is being
implemented using the gtsam library from Georgia Tech.
This implementation is exposed as a shared library.

The main initialization and update functions have been implemented
but there is still some work to do. In particular, the exposed
methods have been designed with C# Interop in mind and they should
be drived from the main framework next.

An appropiate PixelRange factor has been created to work with RGBD
cameras.