Removed slam update debug flags.
As the SlamUpdate method was coded, it was important to get realtime
feedback about the algorithm performance and to separate the effects
of several stages that were used in the filtering approach, so three
flags were set as arguments and could be changed in realtime through
the GUI. These stages don't make sense in the general sense and are
not really useful anymore as the system works correctly.

So they have been "kind of" removed: the main algorithm is now a
clean version with no flags and temporarily and alternate version,
only for PHDNavigator (not the abstract Navigator) can take the
flags, but this alternative is deprecated and will be removed if
it isn't used i the next versions.