Nearest neighbours labeling.
A simple labeling scheme has been implemented: if a measurement
is close to a landmark, it is associated to it; if it's far from
everything, a new landmark is generated.

Though this is simple, it's by no means the best method, since
there's no "screening" of candidate landmarks and they cannot be
removed. Nevertheless it's easy to understand and good enough to
start testing the core functionality of the iSAM2 interface. When
that's complete it should be changed to use a smarter scheme,
e.g. use the already implemented Murty's algorithm to find better
global matches and demand repeated correlative observations before
accepting a new landmark.