iSAM2 update completed.
The update routine is now completed and accounts for labels.

The main structure has been changed from a struct to an object
with a constructor to make it easier to use.
Many variables' scopes have been changed using this persistent
object, where it was convenient.

An incorrect understanding of the iSAM2 internal update call has
been fixed, as the Graph and Estimate variables only represent the
new information at each step, instead of the entire memory.