Resolved motion covariance difference.
The gtsam library uses a 6D noise covariance for the motion model,
unlike the previously 7D noise throughout the program.
This incosistency is now dealt with two internal covariance
variables, one for each one. It should be no big problem.

To relate both variables, the noise can be configured with the 6D
style, which automatically calculates the equivalent 7D covariance
using the jacobian between the coordinate systems.