Graphslam first approximation completed.
All the major components of the graphslam approach have been
implemented and debugged. They seem to work correctly, giving a
similar experience to the PHD navigator. Nevertheless, the
algorithm has trouble with clutter, since its labeling scheme is
quite simple. This should be corrected in the next stage.

It is noted that the algorithm is slow compared to the PHD
implementation, but this could be augmented by the interop barrier.
Further analysis should be performed to compare their speed.