Bugfix: Kinect Delta + Refactoring.
One of Vehicle's constructors used, inappropriately, a constant
defined in KinectVehicle, one of its subclasses. This was coded in
a rush and made little sense. It also produced a bug in the
SimulatedVehicle, where the observable field was incorrect.
The backwards dependency has been removed and the use of the
KinectVehicle constant (Delta) has been moved to the KinectVehicle
class.

Some variables have been renamed.
Some variable uses have been replaced with properties.
A deprecated method has been removed.