Clutter association in iSAM2.
Previously, any new measurement on a uncharted location was added
as a new landmark. Though that works ok if no clutter is introduced,
even a small amount of clutter makes it diverge.
Now, before a new landmark is accepted, it has to be observed in
several consecutive frames first. This generates a much better
map model and is more stable.
It can still become unstable in some situations, though, and it
must be noted that some information is lost "waiting" for the
new landmark to be sufficiently evident.

Next directions are using Murty's algorithm to avoid multiple
measurements to the same landmark, storing the candidate
measurements and using them all when it is an evident landmark.

A bug has been fixed with the transmission of the estimate marginals
from gtsam to the application, since an incorrect index was being
used and a different offset. This meant that the covariance was
meaningless. It has been corrected and now the plots are reasonable.