Removed dummy sidebar.
The sidebar was used only with the KinectVehicle.
If a simulation is used instead, it doesn't make sense to generate
a possibly large file for an empty video and waste screen space.
The Vehicle class has been augmented with a HasSidebar property
that is queried to optimize away any dummy use.

Minor style corrections have been done.