Bugfix: viewer crash on headless output.
Running the system in headless mode generated seemingly correct
output and the viewer showed the same visualization, but at the end
the viewer crashed. This was due to a off-by-one difference between
the sizes of the different recorded data.
The difference in size was due to the way the game time was updated:
in interactive mode it was updated before running the logic code
while in headless it was updated after.
This, along with a bad value for the first "previous time" meant that
in the headless version the first frame was not processed by slam.

Now both system update the time after the logic code (so the first
recorded time is exactly zero) and the data lengths are all the same.

Note however that this probably will need revisiting when the render
and slam framerates differ, as then there would clearly be a difference
in amount of data. Nevertheless, that should be handled in the
Viewer class and is not that urgent, as that case is not interesting
right now.