Gtsam exception handling.
If an error occured in gtsam, e.g. the solution diverged, the program
would just crash and lose all information.
Now it can exit gracefully and save any partial estimates it had
before the situation. This can be used to analyze how critical is
the data association to the method and what's the half-life of a
erroneous system.

Minor style corrections.