Bugfix: GraphCombinatorics Murty test failing.
One of the Murty's algorithm test was failing. It was caused by
an incorrect comparison function being used, the == operator when
it was not defined. Now the operator the same as the Equals method,
which makes sense for the MatrixKey value structure.
That makes the MurtyNode's Equals() to work correctly and then
the test can correctly check that the children set contains the
correct nodes.

Note that no "logic code" has been changed as it was merely a problem
of definitions.