Bugfix: off-by-one frame recording.
Trying to bypass trajectory noise for an experiment, a bug was
noted: the trajectory wasn't being displayed. It turns out this is
caused by the "copy to waypath", as it was being by reference.
This meant that "recorded states" could change later if the vehicle
moved updating but not replacing the state. It so happens that the
real movement is an update, while the noise is a replace; ergo,
when removing the noise, all the recorded states were moved along
and the result was a point at the last location (no visible trajectory).

This result had a minor but important effect in all recordings,
as though less noticeable, it meant all frames were "randomly"
off-by-one (but with the noise of the previous frame, etc.).

Now every recording is performed on a shallow clone of the state
so it is effectively freezed.

Note that a previous decision to update the times at the end of
every frame was wrong, as the most sensible option is to let the
first, pre-loop, state to be zero. This scenario was generating
two states at t=0. Now it starts at zero and __before__ every
update it goes up.