Bugfix: arbitrary landmark numbering.
the iSAM2 processor expected to find landmark association in a
correlative order, i.e. the first measurement should be l0, then l1,
etc.. This would cause an array out-of-bounds write if l24 was
present but l5 was not.
Now it can handle any landmark coming first, but it must be noted
that the used memory is proportional to the max landmark index used,
i.e. if l45 is seen, 46 places will be saved for the rest. This
is the usual logic and it doesn't really make sense to create a
large number if there's no knowledge that smaller indices will
eventually be used by landmarks. Also, this reduces the number of
resize operations required.