Batch processing.
The python post-processing script has been modified to analyze
several runs of the algorithm, showing an average plot of them.
A couple of scripts have been added to streamline the typical
use case to a single shell call.

The recording file for the simulations now contains the scene
data, to simplify the comparisons in the post-processing stage.