Linear assignment for iSAM2.
The system now uses the hungarian method to find the data association
for the measurements. The main algorithm was already implemented
for the PHD weighting scheme so it has been adapted.

It is noted that this change heavily improves the performance of
the algorithm.

A bug has been fixed, where the noise description that was passed
to the C++ isam module was wrong. The gtsam library required the
noise to be expressed in standard deviations, but it was sent as
covariance.

The visualiation has been changed to show 1-sigma ellipse instead
of 5-sigma, because iSAM2 has a lot more variance than the PHD
filter.

Some minor refactoring to comply with guidelines and simplify
the implementation.