Viewer and plotter filtering/smoothing modes.
Now both the viewer and the python plotter can choose between
displaying the "best estimate history" as is, i.e. smoothed in
time, or fixed, i.e. a history of best "current poses".
Both modes can then be easily compared.

Note however, that this is not the same filtering as delivered by
the gtsam library, since the underlying algorithm still solves
the full batch problem. This filtering is what the vehicle's
computer could use in realtime in practice. A stronger filtering
can be generated from the gtsam library where no past data is ever
updated to generate better current data. This is both faster and
much more lightweight but would give worse results.

In the case of PHD, the method is filtering by its very nature so
the filtering mode is "exact". The smoothing version in that case
is choosing the best particle at the end of the simulation and
using its entire history, deciding indirectly on past poses.

Naturally, the smooth case generates smoother curves that change
in time, while filtering can present sudden jumps if a hypothesis
turns out to be far off.

The scene file is no longer required to be specified when using
the viewer as it is embedded in the data file.