Post-processing average change.
When combining different runs for plotting, the system would pick
the first one as a reference for time. This meant that the output
was somewhat arbitrary regarding the order of the arguments
and it could hide potential problems if a shorter run went first.
Now it uses the maximum time and if one of the entries does not
go so far, it is set to infinity so the user can be visually aware
of the situation. Then, he may remove the faulty runs, although
this may not be possible for very unstable settings.