Added RecordVehicle.
A new kind of vehicle has been added, which can use a previously
recorded motion and measurement history. This is similar, but not
the same as the Viewer module, since that module has a lot of
extra visualization capabilities, while it can't interact with the
navigator algorithms. This one only uses a previous vehicle, but
a fresh navigator.
This is very useful for exactly comparing different algorithms as
the exact same run can be used for every one of them.

Some heavy refactoring has been done to simplify the Vehicle
hierarchy. There was some redundant code and some general utility
functions that were extracted from the main classes to avoid clutter.

Some other minor aesthetic changes.