Bugfix: PHD one-particle render.
When RenderAllParticles was set to false, the render would not
contain the complete particle history. In deed, any particle
resampling would kill it. Now it is maintained.
The behavior has changed slightly from the original purpose of this
mode, as it no longer does filtering, but smoothing, just like
the all-particles one. There's a now a better way to control
filter/smooth: the Viewer module.