Modified config file structure.
As there is now a new configuration file, it makes sense to move
the visualization info (mapclip) into it, leaving the scene file to
describe only the vehicle (map.world) and its associated landmarks.

Vehicle can now be written to file.

This info is output in every run, including when using sensor input,
which is important to figure out any special configuration to the
focal length that the sensor may require.

A new SimulatedVehicle constructor has been added to setup the
focal characteristics.

Vehicle now exposes two new read-only properties: Angle and Angle,
which get the axis-angle representation of the orientation.

The landmark setup has been moved from the SimulatedVehicle to
the general Vehicle, as should have happened when the Landmarks
property itself was moved.

Other minor documentation changes.