Bugfix: Kinect focal and screencut.
Focal characteristics weren't being duplicated when cloning a
vehicle. This meant the Navigator would get nonsensical estimates
for the map and trajectory.

The screen cut (scene | sidebar) wasn't working properly on the
Viewer module, as it uses a SimulatedVehicle even if there is a
sidebar to display (which specifically says it doesn't have anything
to output). Now the module overrides this behavior.

Correction in the README file, where a couple of sample were missing
an important parameter.

Two FIXMEs were removed as they were no longer relevant.

Minor refactoring (name changes, 'static'ize a method).