Bugfix: BestParticle and perfect particle.
The PHDNavigator wasn't updating the BestParticle variable
appropriately. This meant that, although the algorithm worked
correctly internally, the output estimate wasn't really the best one
of the lot. Now the BestParticle really points to the best particle
(MAP estimator).
Some commented debug code (perfect particle visualization) was
missing a couple of lines (it was only importing the state from
the real vehicle, but not its history or a good map model).
The testing scenario where the "perfect particle" was added to the
pool of PHD particles wasn't really working, as (incorrectly) only
one particle was being used.
Also, the assumption that the perfect particle would *always* win
was incorrect. Since the system is stochastic, although probable,
there is always the possibility of being surpassed by another
alternative. The test now checks that most of the time the particle
wins, even if it's not always.