Visibility render.
A new mode has been implemented to see which estimated landmarks
can be seen and which cannot. This is to understand the effect
that changing the visilibity model would entail in later commits.

Two minor behavior modifications were required. First, the update
of the Navigator in mapping mode now tracks the vehicle's history
besides its state. This is important later when viewing the recording.
Second, the Viewer uses the history to recreate the state (before,
only the historic trail was being rendered, so no need for the state).
With the correct state, the visibility calculation for the the
estimate is now correct.

The Visible method of the SimulatedVehicle class has been rewritten
to make use of the VisibleM method. The new code is equivalent and
less redundant.

The command line parameters in the project file have been changed,
as they weren't valid anymore.

Other minor refactoring.