Overridable tracking particle behavior.
The system uses SimulatedVehicle particles to track the vehicle
in the PHD navigation mode. Depending on the kind of Vehicle that's
being tracked, special behavior may be warranted, e.g. Kinect's
visibility model may need adjusting.
Dynamic dispatch is now used to select a better SimulatedVehicle
subclass to use for the trackers.
Note that double dispatch would be an appropriate design policy here,
but since it is cumbersome to implement in C# and usually only one
tracker kind is the "correct" one, a simpler design can be effective:
when the incorrect tracker type is generated an exception should be
thrown.