Kinect visibility model.
A new Kinect tracking vehicle has been added, which changes its
visibility model. Besides having to be in the frame and in the
appropiate range, the landmark now also has to be in front of
the depth map to be deemed visible. This follows from occlusion:
if something has been detected in front of the landmark (even if
it is artificial), there's no way it can be detected.
Note that if the depth map indicates an "fault reading" around the
landmark, it also shouldn't be considered visible. These readings
are tagged as zero here so they are always in front of anything,
ergo no need for special treatment.

The TrackClone methods in the KinectVehicle now return this kind
of tracker.

The VisibleM method in the Vehicle class is now virtual.

The default max number of map components has been raised to be used
in bigger maps.