Kinect depth rendering.
The kinect vehicle now renders its depthmap into a grid in the
visualization 3D space.

Minor modification to KinectVehicle's default initial pose, to
be aligned to the viewport.

The visibility frustum now can be "in front" of other objects to
help the user understand the vehicle pointing direction.

A minor error has been corrected in the default constructor of
Vehicle, since it used the axis {0, 0, 0}, which is not valid.

The keyboard up/down rotation was reversed.