Bugfix: strange correlation in measurements between frames.
As the vehicle moved, some of the measurements seemed to get "stuck"
at a certain point in measurement space. This was due to the
landmark-to-measurement function not giving a sign to the range
variable. Now it does, so landmarks that are behind the camera have
negative ranges, are correctly ordered and can be automatically
filtered out.
This wasn't noted because previously the Visible() function filtered
out these landmarks directly through their local Z axis component.
As this function has been reimplemented reusing other functions,
the special behavior was lost.