Separation between vehicle update and odometry.
There are two kinds of noise in a vehicle update: odometry readings
are corrupted with errors, and as part of the abstract model of the
vehicle, a motion noise is assumed and particles are sampled from it.

Until now both sources were treated the same and the difference was
only made as to how the Vehicle object was being used (the reference
vehicle's noise would refer to odometry and the tracker vehicles
were from the particle filter implementation).

Now there is a harder separation of behavior: the Update function
only changes the vehicle state, deterministically. A new ReadOdometry
function obtains a corrupted cumulative reading. A new function
UpdateNoisy changes the state and adds model noise. This way, every
use is explicit and can't be confused.

As a side effect, odometry corruption can now also be simulated for
the KinectVehicle.

To enforce even more the distinction between tracking and simulation
vehicles, a new class TrackVehicle has been derived from
SimulatedVehicle. Structurally it has no difference, but noise can
only be added to descendants of this TrackVehicle class (they implement
UpdateNoisy), so they should be used for particle tracking.