New plot mode: average error.
Although the two current plot methods (filter and smooth) give a
good sense of how the algorithm performs, there is room for
improvement.
Smoothing has a hard time showing the change in performance as the
navigator moves in time, since it uses the last estimate so all
information has already been fused into the result. This means that
"loop closures" are hard to see because when they occur, they change
the past, hiding themselves from the final result.
Filtering is different, since it shows information as it is obtained
in time, so closures can have a big impact in the current estimate.
The problem is that this plot is too noisy, as the best estimate can
change rapidly from frame to frame (low redundancy in information about
new events).
Loop closures have their biggest impact on the trajectory as a whole
so measuring the whole trajectory error as time goes by is interesting.
The new TimedAverage mode accomplish this by averaging the errors
of the trajectory in time. The plot is smoother than the alternatives
and clearly shows trends in the error as a result of drift and the
moments when the navigator closes a loop, instantaneously decreasing
the average trajectory error.

Exceptions in the subprocesses are now shown in the console instead
of obscure general messages.

If files are not found for the map or trajectory estimates, their
plot is just bypassed instead of crashing the program.