Stop RecordVehicle at end of recording.
Currently, RecordVehicle would use the prerecorded files to recreate
the scene, but after the files were over, it got stuck in the last
frame. This works if its used interactively, as the user may shut
down the application manually, but doesn't interact well with
the automated scripts.
A new flag is set in Vehicle if it think it has finished its
mission and the Simulator complies with it, finishing the application
when the vehicle has nothing left to do, e.g. the prerecorded files
are over.

Additionally, some errors have been corrected in the recordings,
where a RecordVehicle would not record the same output as its input
(an extra zero state would be added at time zero), which can give
inconsistent results between different runs.
It was due to an off-by-one in the indexing used in the class. It
has been fixed and the output is now exactly the same as the input.