Bugfix: motion covariance scale.
The iSAM2 code was missing an important scaling factor when passing
the motion noise covariance, which meant it was 30 times what it should.
This changes somewhat the resulting estimates means, but not that much.
The landmark covariances, though, are much smaller and comparable
to the PHD ones.

Since the covariances are now smaller, the gaussian ellipses were
boosted to 5-sigma again.