Lie covariance.
The motion covariance that was being used in iSAM seemed to be
incorrect, since it was wrt. to the angles yaw-pitch-roll and the
library uses a Lie algebra representation.
The correct covariance is now calculated in the Lie space, but it
turns out its basically the same as the angle, so the behavior of
the system hasn't changed. Anyway, at least now the implementation
is known to be correct.