Auto-input mapping/SLAM switch.
The auto-input file now can specify that the navigator should move
to mapping mode or to SLAM mode. This way, the simulation can start
with a couple of seconds of known trajectory, which is useful for
removing the initial uncertainty, when no map is known, so as to
fix the euclidean invariance. This means that the calculated errors
better reflect the quality of the algorithm.

A minor bugfix was done on the reset() function of PHDNavigator,
which would not carry the trajectory of the tracked particle,
which would then remove part of the estimate on the plots giving
spurious errors.